Finished a few propositions
Finished NeighborLR(left and right) and neighborUD(up and down) as well as finish writing to check if the nieghbor is facing the right direction(eg. if the neighbor on the left is facing east and the neighbor on the right is facing west then add it to c and append it). I have also made all values in PIPE_TYPE variable an array so PIPE_TYPE is a 2d array.